rustup. *phew*

+++ examples/simple.rs
@@ -1,3 +1,5 @@
+#![allow(unstable)]
+
 extern crate csv;
 
 use std::path::Path;
@@ -7,7 +9,7 @@ fn main() {
     let mut rdr = csv::Reader::from_file(fp);
 
     for record in rdr.decode() {
-        let (s1, s2, dist): (String, String, uint) = record.unwrap();
+        let (s1, s2, dist): (String, String, usize) = record.unwrap();
         println!("({}, {}): {}", s1, s2, dist);
     }
 }
